extensions/cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ccfddl"
version = "0.1.0"
description = "CLI tool for CCFDDL - Conference Deadline Tracker"
readme = "readme.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "0x4f5da2", email = "me@4f5da2.com"}
]
dependencies = [
    "tabulate",
    "pyyaml",
    "requests",
    "termcolor",
    "icalendar",
    "xlin",
]

[project.optional-dependencies]
dev = [
    "mypy",
    "pytest>=7.0",
    "pytest-cov",
]

[project.scripts]
ccfddl = "ccfddl.__main__:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["icalendar.*", "xlin.*", "termcolor.*"]
ignore_missing_imports = true